Fix autoconf issues with Xcode 12
On recent versions of Xcode, calling an implicitly defined function
is now an error by default (matching C++'s behavior)  Since most
compilers have warned about this for ages it's not usually a problem.
However it is very common that autoconf scripts will generate warnings
that nobody looks at.

One of the easist ways to trigger this is to call exit()
without #include <stdlib.h>.  However the simplest fix is usually
to just use "return" instead.

Also switch to modern AC_MSG_ERROR instead of AC_ERROR and fix a
const issue that was breaking sys_siglist[] detection for me.